# backend/docs/swagger.yaml
basePath: /api
definitions:
  domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      onboarding_completed:
        type: boolean
    type: object
  handler.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.RecommendRequest:
    properties:
      contextual:
        type: boolean
      direction:
        enum:
        - book_to_movie
        - book_to_book
        - movie_to_movie
        - movie_to_book
        type: string
      exclude_ids:
        items:
          type: string
        type: array
      selected_ids:
        items:
          type: string
        type: array
      weights:
        additionalProperties:
          format: float64
          type: number
        type: object
    required:
    - direction
    - selected_ids
    type: object
  handler.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  handler.recordViewRequest:
    properties:
      item_id:
        type: string
      item_type:
        enum:
        - book
        - movie
        type: string
    required:
    - item_id
    - item_type
    type: object
  redis.RecommendationResult:
    properties:
      contextual:
        type: boolean
      created_at:
        type: integer
      error:
        type: string
      movies:
        items:
          type: string
        type: array
      status:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Cross-media recommendation system API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Recly API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: access_token, refresh_token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "409":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Register a new user
      tags:
      - auth
  /book/{id}/like:
    delete:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Remove book from library
      tags:
      - library
    post:
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Add book to library
      tags:
      - library
  /genres:
    get:
      parameters:
      - default: all
        description: Item type (book/movie/all)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: genres
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get genres
      tags:
      - search
  /items/batch:
    get:
      parameters:
      - description: Comma-separated IDs
        in: query
        name: ids
        required: true
        type: string
      - default: all
        description: Item type (book/movie/all)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: items
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get multiple items by IDs
      tags:
      - search
  /movie/{id}/like:
    delete:
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Remove movie from library
      tags:
      - library
    post:
      parameters:
      - description: Movie ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Add movie to library
      tags:
      - library
  /recommend:
    post:
      consumes:
      - application/json
      parameters:
      - description: Recommendation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RecommendRequest'
      produces:
      - application/json
      responses:
        "202":
          description: task_id, status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Request recommendations
      tags:
      - recommendations
  /result/{taskId}:
    get:
      parameters:
      - description: Task ID
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/redis.RecommendationResult'
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get recommendation result
      tags:
      - recommendations
  /search:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - default: all
        description: Item type (book/movie/all)
        in: query
        name: type
        type: string
      - description: Genre filter
        in: query
        name: genre
        type: string
      - description: 'Sort field: ratings_count:desc, vote_count:desc'
        in: query
        name: sort
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: results
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Search items
      tags:
      - search
  /user/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Avatar image (jpg, png, webp)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: avatar_url
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Upload avatar
      tags:
      - user
  /user/library/books:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: books
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get liked books
      tags:
      - library
  /user/library/movies:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: movies
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get liked movies
      tags:
      - library
  /user/onboarding/complete:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Complete onboarding
      tags:
      - user
  /user/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: user profile
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: user not found
          schema:
            additionalProperties: true
            type: object
      summary: Get user profile
      tags:
      - user
  /user/recommendations/history:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: history
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get recommendation history
      tags:
      - recommendations
  /user/saved-items:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: saved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get saved items
      tags:
      - saved
    post:
      consumes:
      - application/json
      parameters:
      - description: Item data (item_type, item_id)
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: saved
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Save item
      tags:
      - saved
  /user/saved-items/{id}:
    delete:
      parameters:
      - description: Saved item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: message
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Delete saved item
      tags:
      - saved
  /user/view:
    post:
      consumes:
      - application/json
      parameters:
      - description: View data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.recordViewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: status
          schema:
            additionalProperties: true
            type: object
        "400":
          description: error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Record view
      tags:
      - views
  /user/views:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: views
          schema:
            additionalProperties: true
            type: object
        "401":
          description: unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: error
          schema:
            additionalProperties: true
            type: object
      summary: Get recent views
      tags:
      - views
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"